app_embalagem/requirements.txt
PySide6==6.8.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-barcode==0.15.1
Pillow==11.0.0
